### 2024-05-26 #6792 @editorialbot
[PRE REVIEW]: HoverFast: an accurate, high-throughput, clinically deployable nuclear segmentation tool for brightfield digital pathology images
**Submitting author:** @petroslk (<a href="http://orcid.org/0009-0005-2015-6795">Petros Liakopoulos</a>)
**Repository:** https://github.com/choosehappy/HoverFast
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
