### 2024-05-28 #6792 @Kevin-Mattheus-Moerman
@petroslk thanks for this submission. I have briefly reviewed the above analysis, the paper, and your repository.

* Because of the relatively small size and narrow functionality of this work I have just triggered a scope review by the editorial board to determine if this work conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

Required changes before we can proceed:

- [ ] Use an [OSI approved open source license](https://opensource.org/license).
- [ ] Add contributing guidelines to your project.
- [ ] Add detailed documentation, the README information seems very basic.
- [ ] Add testing instructions and link to them in the README. Automated testing is preferred, is that available?
- [ ] Try to address the above DOI issues. You can call `@editorialbot check references` to check them again once you've amended your .bib file.
- [ ] In your affiliations please spell out all country names in full (do not use country acronyms).
