### 2024-05-29 402d64f
code reformatting, contibutions, documentation and improved testing
