+++ .readthedocs.yaml
@@ -1,4 +1,4 @@
-version: "1"
+version: "2"
 
 build:
   os: "ubuntu-22.04"
