+++ docs/requirements.txt
@@ -2,3 +2,4 @@ sphinx
 sphinx-autobuild
 sphinx-rtd-theme
 sphinx_design
+furo
