### 2024-05-29 a0d1667
reformat index.rst
