### 2024-05-29 2512a77
fix toctree max depth
