+++ docs/source/conf.py
@@ -9,7 +9,7 @@
 project = 'HoverFast'
 copyright = '2024, Petros Liakopoulos, Julien Massonnet, Andrew Janowczyk'
 author = 'Petros Liakopoulos, Julien Massonnet, Andrew Janowczyk'
-release = '1.0'
+release = '1.0.0'
 version = '1.0.0'
 
 # -- General configuration ---------------------------------------------------
