### 2024-05-29 a9e2f9e
remove functions from readthedocs
