+++ docs/source/infer_wsi.rst
@@ -163,14 +163,3 @@ Explanation:
 
 
 
-Functions
----------
-
-The `infer_wsi` module contains the following functions:
-
-.. dropdown:: Click to show/hide functions
-
-    .. automodule:: hoverfast.utils_wsi
-        :members:
-        :undoc-members:
-        :show-inheritance:
