+++ docs/source/training.rst
@@ -188,29 +188,3 @@ Explanation:
 - Dataset Path: Directory that contains the pytables.
 - Epochs: Set the number of epochs to 200.
 - Output directory: Directory to save TensorBoard logs and trained model.
-
-Functions
----------
-
-The `train` module contains the following functions:
-
-.. dropdown:: Click to show/hide functions
-
-    .. automodule:: hoverfast.training_utils
-        :members:
-        :undoc-members:
-        :show-inheritance:
-
-.. dropdown:: Click to show/hide functions
-
-    .. automodule:: hoverfast.augment
-        :members:
-        :undoc-members:
-        :show-inheritance:
-
-.. dropdown:: Click to show/hide functions
-
-    .. automodule:: hoverfast.hoverfast
-        :members:
-        :undoc-members:
-        :show-inheritance:
