### 2024-05-29 5cf937c
Update README.md with full documentation
