### 2024-05-30 35e9908
address missing DOIs and add citation info
