+++ paper/paper.bib
@@ -25,6 +25,7 @@
   number={1},
   pages={100--106},
   year={2021},
+  doi={10.1038/s41592-020-01018-x},
   publisher={Nature Publishing Group US New York}}
 
 @article{graham2019hover,
@@ -34,6 +35,7 @@
   volume={58},
   pages={101563},
   year={2019},
+  doi={10.1016/j.media.2019.101563},
   publisher={Elsevier}}
 
 @article{bankhead2017qupath,
@@ -44,6 +46,7 @@
   number={1},
   pages={1--7},
   year={2017},
+  doi={10.1038/s41598-017-17204-5},
   publisher={Nature Publishing Group}}
 
 @article{su2021msu,
@@ -53,6 +56,7 @@
   volume={12},
   pages={639930},
   year={2021},
+  doi={10.3389/fgene.2021.639930},
   publisher={Frontiers Media SA}}
 
 @article{paszke2017automatic,
@@ -66,7 +70,9 @@
   journal={Journal of Machine Learning Research}, 
   volume={12}, 
   pages={2825--2830}, 
-  year={2011}}
+  year={2011}},
+  url={http://jmlr.org/papers/v12/pedregosa11a.html},
+  }
 
 @article{opencv_library,
   author = {Bradski, G.},
@@ -86,6 +92,7 @@
   number={10},
   pages={1113--1120},
   year={2013},
+  doi={10.1038/ng.2764}
   publisher={Nature Publishing Group}}
 
 @article{hinton2015distilling,  
@@ -97,7 +104,9 @@
 @article{hu2022teacher,
   title={Teacher-student architecture for knowledge learning: A survey},  
   author={Hu, Chengming and Li, Xuan and Liu, Dan and Chen, Xi and Wang, Ju and Liu, Xue},
-  journal={arXiv preprint arXiv:2210.17332},  year={2022}}
+  journal={arXiv preprint arXiv:2210.17332},
+  year={2022},
+  }
 
 @article{janowczyk2019histoqc,
   title={HistoQC: an open-source quality control tool for digital pathology slides},
@@ -106,7 +115,8 @@
   volume={3},
   pages={1--7},
   year={2019},
-  publisher={American Society of Clinical Oncology}
+  doi={10.1200/CCI.18.00157},
+  publisher={American Society of Clinical Oncology},
   }
 
 
