+++ paper/paper.bib
@@ -92,8 +92,9 @@
   number={10},
   pages={1113--1120},
   year={2013},
-  doi={10.1038/ng.2764}
-  publisher={Nature Publishing Group}}
+  doi={10.1038/ng.2764},
+  publisher={Nature Publishing Group}
+  }
 
 @article{hinton2015distilling,  
   title={Distilling the knowledge in a neural network},  
