### 2024-05-30 #6792 @petroslk
Dear @Kevin-Mattheus-Moerman,

Thank you for your reply and for considering our work. As requested I have addressed your feedback. I have:
- Changed the License to BSD-3-Clause (OSI approved)
- Added contributing guidelines
- Added more in depth documentation with a separate [readthedocs page](https://hoverfast.readthedocs.io/en/latest/).
- Added testing instructions on the readme and on readthedocs. Unfortunately, since the tool is GPU based, I have not yet been able to implement automated testing and run those locally.
- I have included correct DOIs for all references that had them.
- I have corrected the affiliation acronyms and included full Country names.

Let me know if there is anything missing.

Considering we had to reset the GitHub repository for the submission, most of the commit history is not visible. We also decided to provide the following statement for the editorial board to consider:

---

### Statement of Substantial Scholarly Effort

We, the coauthors, have invested upwards of six months optimizing and iterating over the codebase to achieve the required performance for HoverFast. Our primary focus was to maintain a compact and efficient codebase while ensuring high accuracy and throughput, filling an important niche in the field of digital pathology. Due to the nature of our clinical work and the need to provide a clean repository for this submission, we reset the GitHub repository, which resulted in the loss of the commit history that reflects our extensive development efforts. However, we can attest that the work involved:
- Detailed performance optimizations and code refactoring.
- Comprehensive testing and validation on various datasets.
- Collaborative development efforts among multiple authors, including medical doctors.
- Iterative enhancements based on user feedback and practical deployment scenarios.

HoverFast is designed to be feature-complete and adheres to common community standards for Python-based tools. It addresses a critical need for a high-throughput, clinically deployable nuclear segmentation tool for brightfield digital pathology images, representing a substantial scholarly effort and a significant contribution to the available open-source software for digital pathology. We believe that HoverFast meets and exceeds the minimum contribution threshold set by JOSS and will be a valuable tool for researchers and practitioners in the field.

---

Please let me know if you are in need of any additional information or any feedback to provide! Thank you once again for considering our submission.
