### 2024-05-30 #6792 @Kevin-Mattheus-Moerman
@petroslk below I have two minor points in relation to the paper:

- [ ] Please add `United Kingdom` to your affiliation
- [ ] Check if `generalizable` should be `generalisable` to be consistent with the rest of the text which uses British English
