+++ tests/test_roi_inference.py
@@ -21,7 +21,13 @@ def single_roi_dir(tmp_path_factory):
     yield pth
 
 
-def test_infer_roi(tmp_path):
+def test_infer_roi_he(tmp_path):
     rv, out = getstatusoutput(f"{PRG} infer_roi {ABS_TILE_PATH}/*.png -m {ABS_PATH_MODEL} -o {tmp_path}")
     assert rv == 0
     assert _filenames_in(tmp_path) == _filenames_in(tmp_path).union(["json", "label_mask", "overlay"])
+
+def test_infer_roi_ihc(tmp_path):
+    rv, out = getstatusoutput(f"{PRG} infer_roi {ABS_TILE_PATH}/*.png -m {ABS_PATH_MODEL} -o {tmp_path} -a ihc_dab")
+    assert rv == 0
+    assert _filenames_in(tmp_path) == _filenames_in(tmp_path).union(["json", "label_mask", "overlay"])
+
