### 2024-06-24 #6792 @Kevin-Mattheus-Moerman
@petroslk I've just removed the `query-scope` label as I'm pleased to confirm that this submission has passed the initial scope review process. We will now be looking for a handling editor.
