### 2024-07-08 #6792 @sappelhoff
Hi @petroslk 👋  I will be editing your submission to JOSS. My first question to you is: Can you recommend potentially suitable reviewers from your larger academic network? If yes, please let me know and make sure that they would fulfill the [JOSS COI guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).

If not, please let me know, too -- then I will try to assemble a team via other databases, or potentially the [automatic suggestions](https://github.com/openjournals/joss-reviews/issues/6792#issuecomment-2184972357).
