### 2024-07-16 #6792 @PingjunChen
@sappelhoff I am delighted to review HoverFast for JOSS.
