### 2024-07-19 1d2e46c
Add documentation for IHC DAB support
