### 2024-07-28 choosehappy/HoverFast#2 @PingjunChen
[JOSS Review] - PingjunChen
Hey,

I am reviewing the checklist https://github.com/openjournals/joss-reviews/issues/7022#issuecomment-2243749071. HoverFast is in good shape overall. Two minor issues:

1. The command line for training the model using docker/singularity in the "README/Training/Train model" part, "-l training_metrics" should be "-o training_metrics".
2. "scikit-learn’s [@scikit-learn] regionprops and watershed" inside [paper/paper.md](https://github.com/choosehappy/HoverFast/tree/main/paper) should be scikit-image.
