### 2024-07-30 33cebc9
Made modification for joss review (issue #2)
