### 2024-07-30 choosehappy/HoverFast#2 @JulienMassonnet
Hi @PingjunChen,

Thank you very much for your feedback. I have made the necessary modifications and recompiled the paper to verify if it's working. Please let me know if there are any other points you would like to address.
