### 2024-07-30 #7022 @PingjunChen
@sappelhoff The authors have addressed my minor concerns. I believe HoverFast would be a quite useful tool for the computational pathology community, particularly considering its speedy inference compared to Hover-Net. I recommend its acceptance for publication.
