### 2024-09-23 choosehappy/HoverFast#4 @NetoPedro
[JOSS Review] - NetoPedro
Hi,

I am reviewing the checklist https://github.com/openjournals/joss-reviews/issues/7022#issuecomment-2245007276.

Overall the paper looks good and the improvements are outstanding with regards to the computational cost. I have single question (or two but in the same direction). What is the impact of KD here? How different would be the results if HoverFast was trained just with the original labels?

Thanks.
