### 2024-09-23 #7022 @sappelhoff
@petroslk please let me know once you finished your items from the list above: https://github.com/openjournals/joss-reviews/issues/7022#issuecomment-2368183621

I will then work on my final items, and then it's (hopefully) finally time to ship it!
