### 2024-09-25 d86250c
minor bug fix infer_roi
