### 2024-09-25 #7022 @petroslk
Hey @sappelhoff

As requested, the release has been archived on Zenodo (DOI: https://doi.org/10.5281/zenodo.13838349 ). I have also checked author names and Orcid IDs for those that provided them.
This release is of version 1.0.1 which includes reviewer comments.

Let me know if there anything else I might have missed.

Thanks a lot!
