### 2024-09-25 #7022 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
