### 2024-09-26 #7022 @Kevin-Mattheus-Moerman
@petroslk as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [x] Software license tag listed here matches a tagged release
